made changes here to get stuff into the ROOT Files I believe.  not sure on the spacifics or I would have added them in here one by one.
However Wouter asked me to make the commit, so this is all I got.
Clearly more understanding needs to happen.

One GREAT thing though is Wouter and I have added the ablity to make
units in the ROOT tree so that one will *know* exactly what the units
are of what they are plotting!! How Cool is that?!