Addition of a new HMolPol macro.
This macro will output a ROOT simulation file to look at the hits that
come out of the simulation.  As of 02-08-2015 there is a problem with
the hits getting recorded. 

It includes the ablity to turn on and off all the physics processes in
order to see if one of these is the culperit. 

NOTICE: The ROOT file name has places that should get filled with the
date and process(es) on or off.