Added in format and correct output.
- Changed the output of the header file to ./include/ rather that the
defined ./build/include/ from the remoll version.
- Changed the name of the output to HMolPolVersion.hh, that which is
already used in the code by the HMolPol file for run information.
- Created the output to the HMolPolVersion.hh to have the correct form,
following the example from Qweak. (QweakSimVersion.hh &&
QweakSimUserInformation.hh and .cc)


Note: The file outputs, but there is some chnages that have to be made
because of the way Python reads in command. I am working on changeing
that. However I need to commit as I will not be working on this computer
today. Stand By for these changes.