Added a new variable to the ROOT tree and modified one.
Modifed names fPhiLab to fPhiLab1 and added in fPhiLab2.

In the code for PrimaryActionGenerator, there is only one theta_com and
phi_com (Center of Mass angles) because of the trig identies (-
cos(theta) = cos(theta + 180) and -sin(theta) = sin(theta + 180) for
degrees) we can add in the needed minus signs (this time they are not
magic ones) which means that there is no reason to define a second
Center of Mass theta and phi angles.

we know that:
fPhiLab1 = phi_com;
fPhilab2 = (phi_com + pi <2*pi) ? (phi_com = pi) : (phi_com - pi);

This defintition of the if statment for fPhiLab2 is such that the range
of fPhiLab1 and fPhiLab2 is that of standard phi range for spherical
coordinates or 0 to 2*pi (0 to 180 deg).