Added new variables to the Primary Branch
I added the following variables to the primary branch:
 - fThetaCenterOfMass1 - center of mass theta for particle 1, theta_com
 - fThetaCenterOfMass2 - center of mass theta for particle 2, 180 -
theta_com
 - fPhiCenterOfMass1 - center of mass phi for particle 1, phi_com
 - fPhiCenterOfMass2 - center of mass phi for particle 2, phi_com + 180
 
Also I made the interaction position variables into a TVector3 and the
interaction mometum for particle 1 and 2 variables into a TVector3. This
is to make things easier as a vector has functions built in like
magnitude. 

While doing this I corrected the Oops where the momentum for particle 1
and particle 2 was stored in the same variable, particle 2 information
overwriting the particle 1 information.

This means that the tree structure has changed enough the root files
made with this commit or commits afterwords will not be able to be read
wtih versions of hmolpolroot from commits this. Also rootfiles made from
commits before this will not be able to be read by this version of
hmolpolroot.