Addition of Cross Sections
I added in the cross section calculation. The only part that is outputed
to the tree is the one we care about, the CM cross section this is from
the High Percison paper (reference in code).

Also there is the Qweak and remoll cross sections as of this date and a
disciption of what they are. As well as conversion from CM frame to lab
frame for both the relativistic and non-relativistic cases. Of course we
want the former, but the latter is what one finds most when looking for
this. References for these conversions are also stated. Howver these are
not put into the tree as like stated before we care about the CM cross
section. 

Also all of these with the excetion of the copy and paste from Qweak are
stored in the units of mm^2. This is due to interal workings of Geant4.
So divide by the units you want. The way Qweak is done microbarns are
stored as mm^2 if you were to output it to the Root file in HMolPol.
(this may only matter when you have a units branch like I do, I am not
sure. I am only talking about HMolPol here)