Addition of units for cross section
Added in units for cross section, units of barns, microbarns.

Note Geant4 caclulates cross sections in mm^2. So these are needed for
getting things in barns, etc. And well that is how we want it.

Remember that the cross section is the differental one so really the
untis are area/sr. But this isn't something Geant4 will complie with.