Cross section Root tree variable
I forget to add this in with the cross section commit, this commit has
to do with the storing of the cross section in the ROOT file. I changed
the name to fCrossSectionCM. One big reason is because this makes sure
that there is no question on what cross section it is, CM or Lab.