Changed material of the cell around H gas
I changed the material for the cell around the H target. According to
Appendix A, pg 33 the material is copper with a:
- length of 0.4 m, which contradicts pg 7 but what I used
- diameter of 0.02 m (inner there is no outer I can find)
- Temp or 0.3K

The specific heat and conductance is noted but I am unable to put that
into the GDML and it seems Geant4 in general will not take that
information.