Addition of a Primary Condition Branch
I added in a branch that will store all the primary condtions of the
simulation. Right now that consist of:
- Beam Energy
- Raster size in x and y
- Min and Man theta and phi values tossed over
- the solid angle which is tossed over
- the change in theta and phi that is tossed over
- the difference in the cos(theta) that is tossed over

These values are passed from the primary action generator where they are
used and set to the PrimaryConditions where they are stored in the ROOT
tree. This required adding of this branch in to the Root tree via the
HMolPolAnalysis class.

Moving forward other global information will be stored here in the ROOT
file from the simulations like field strenths, dimenstion and such.