Macros for calculating the Cross Section and writing out the intial conditions.
CrossSection.C: is a macro that will calculate the cross section of a
giving simulation for events that make it to the detector. This function
works for both simulations ran on the dummy detector and the LeftRight
detector. The detector just has to be specified.

ExtractInitalConditions.h: macro that must be included in a different
one, like the run info one. It will write out the initial conditions of
the simulation to a file. This must be included after one defines the
output prefix. Also it must be called when one is looping through the
tree once you have an entry, it needs to be past the units and
PrimaryConditions branch. If you call it before you have an entry then
you will only get the values which are constants out and nothing else.
But it only needs to be called once, a nice if(ientry == 1) will do the
trick once you have started looping through the tree.