Geometry for the Hall A magnetic optics
Files for the Hall A magentic optics.
Quads 1 and 4 have a file as they are the same type.
Quads 2 and 3 have a file as they are the same type.
Dipole has its own file.

All of the magnetics are placed in there relative locations in the
HallA_Optics.gdml file.

The "new" mothervolume file includes the geometry for the optics in it
with the dummy detector. 

All locations have been found via the hall A surveys, and the sizes have
come frim other hall A magnetics information.