Added Radial_Distribution.C
I added in a script that makes hit count distributions vs distance away from the center of the dummy detector for primary electrons.  For each primary cut, it makes a distribution for polar angle ranges of 0 to pi/6, pi/6 to pi/3, and pi/3 to pi/2 on the face of the dummy detector.  I suspect any more would be redundant for now.