Added new materials
I added new materials.

In nist_materials.txt, I added
G4_LEAD_GLASS
G4_SILICON_DIOXIDE

I created a Lead glass material in materials.gdml which is modeled
closely after the Schott SF-5 Lead Glass. Which is one that is commonly
used. However I had to get the materials it is made of which I borrowed
from a Fermi Lab simulation. Which is noted in the file.

Note: The density of the two lead glass materials are different. Among others properties, it was unsure as to how to add them in to GDML for the SF-5 material, but this gives two choices going forward, until something is decided or a new material is put in.