Changes to make the magnetic field tags better supported. Now it creates a map from tag type to value. Also rotation and translation for quadrupoles are supported now, which may be required for the Hall A optics to work.
Rotation: it picks the rotation of the magnetic volume and applies any additional
rotation specified by Euler angles in degrees in the GDML file.

Translation: it picks the center of the magnetic volume as the center of the
quadrupole field and applies any additional shift as specified by a vector
in meters in the GDML file.