Added the ability to easily change the Z generator location
Added a command to changed the Z generator location in the messenger.
Which is then stored in the Primary Condition branch and used in the
primary generator.

The standard in the primary generator has a Z generator range of -10 cm
to 10 cm.

Added the commands to myRun.mac as an example.