Adding the Analyzing Power to the Simulation and ROOT tree
* Added a new variable to the fPrimary Branch: fA_zz which is the
Analyzing power for that event 
  - HMolPolEventPrimary.hh
  - HMolPolPrimaryGeneratorAction.cc
  
* Changed the analysis file CrossSection.C to CrossSection_Analysis.C so
now it not only calculates the cross section but also the Analyzing
Power. And writes that to the Text file. 
  - Note: The LEFTRIGHT detector part, will have to be check when
needed. The LEFTRIGHT detector is not being use and therefore was
written, but not tested. It should work but there might be a bug there.